+++ src/lark-meeting/lark-meeting.module.ts
@@ -2,7 +2,7 @@
  * @Author: 杨仕明 shiming.y@qq.com
  * @Date: 2025-11-22 23:46:35
  * @LastEditors: 杨仕明 shiming.y@qq.com
- * @LastEditTime: 2025-11-23 18:26:43
+ * @LastEditTime: 2025-11-23 18:35:38
  * @FilePath: /lulab_backend/src/lark-meeting/lark-meeting.module.ts
  * @Description:
  *
@@ -12,10 +12,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { LarkWebhookController } from './controllers/webhook.controller';
-import { LarkEventWsService } from './service/lark-event-ws.service';
+import { LarkEventWsService, LarkMeetingService } from './service';
 import { LarkModule } from '../integrations/lark/lark.module';
-import { LarkMeetingDetailService } from './service/lark-meeting-detail.service';
-import { LarkMeetingService } from './service/lark-meeting.service';
 import { LarkEventProcessor } from './queue/lark-event.processor';
 import { BullModule } from '@nestjs/bullmq';
 import { larkConfig } from '@/configs/lark.config';
@@ -27,12 +25,7 @@ import { larkConfig } from '@/configs/lark.config';
     BullModule.registerQueue({ name: 'lark-events' }),
   ],
   controllers: [LarkWebhookController],
-  providers: [
-    LarkEventWsService,
-    LarkMeetingDetailService,
-    LarkMeetingService,
-    LarkEventProcessor,
-  ],
-  exports: [LarkMeetingDetailService, LarkMeetingService],
+  providers: [LarkEventWsService, LarkMeetingService, LarkEventProcessor],
+  exports: [LarkMeetingService],
 })
 export class LarkMeetingModule {}
