+++ src/lark-meeting/service/index.ts
@@ -1,0 +1,2 @@
+export * from './lark-event-ws.service';
+export * from './lark-meeting.service';
